+++ bitarray/sparse_bitarray_test.go
@@ -121,10 +121,10 @@ func BenchmarkGetSetCompressedBits(b *testing.B) {
 
 func TestCompressedCount(t *testing.T) {
 	ba := newSparseBitArray()
-	assert.Equal(t, uint64(0), ba.Count())
+	assert.Equal(t, 0, ba.Count())
 
 	require.NoError(t, ba.SetBit(0))
-	assert.Equal(t, uint64(1), ba.Count())
+	assert.Equal(t, 1, ba.Count())
 
 	require.NoError(t, ba.SetBit(40))
 	require.NoError(t, ba.SetBit(64))
@@ -132,13 +132,13 @@ func TestCompressedCount(t *testing.T) {
 	require.NoError(t, ba.SetBit(200))
 	require.NoError(t, ba.SetBit(469))
 	require.NoError(t, ba.SetBit(500))
-	assert.Equal(t, uint64(7), ba.Count())
+	assert.Equal(t, 7, ba.Count())
 
 	require.NoError(t, ba.ClearBit(200))
-	assert.Equal(t, uint64(6), ba.Count())
+	assert.Equal(t, 6, ba.Count())
 
 	ba.Reset()
-	assert.Equal(t, uint64(0), ba.Count())
+	assert.Equal(t, 0, ba.Count())
 }
 
 func TestClearCompressedBit(t *testing.T) {
